refactor(materializer): extract public types from canonical graph materializer (file #3 start)
First increment of the canonical agent-panel-graph-materializer decomposition.
Lift the public type surface (AGENT_PANEL_SCENE_TEXT_LIMITS + the
Header/Materializer Input + ReadModel interfaces) into a -types module; the
materializer re-exports them (zero consumer churn). GOD-safe: pure type/const
relocation, no canonical logic touched. Characterization net 74/74 green, check 0.
