+++ packages/desktop/src/lib/acp/session-state/agent-panel-graph-materializer.ts
@@ -38,41 +38,24 @@ import {
 import { getInteractionSnapshotArrayPatch } from "./interaction-snapshot-array-patch.js";
 import { getOperationSnapshotArrayPatch } from "./operation-snapshot-array-patch.js";
 import { getTranscriptEntryArrayPatch } from "./transcript-entry-array-patch.js";
+import {
+	AGENT_PANEL_SCENE_TEXT_LIMITS,
+	type AgentPanelGraphHeaderInput,
+	type AgentPanelGraphMaterializerInput,
+	type AgentPanelGraphMaterializerReadModel,
+} from "./agent-panel-graph-materializer-types.js";
+
+// Re-export the public type surface (now owned by the -types module) so the
+// materializer's existing consumers keep importing it from here.
+export {
+	AGENT_PANEL_SCENE_TEXT_LIMITS,
+	type AgentPanelGraphHeaderInput,
+	type AgentPanelGraphMaterializerInput,
+	type AgentPanelGraphMaterializerReadModel,
+};
 
 const TRUNCATION_SUFFIX = "\n[truncated]";
 const UNRESOLVED_TOOL_DIAGNOSTIC_SAMPLE_LIMIT = 40;
-
-export const AGENT_PANEL_SCENE_TEXT_LIMITS = {
-	output: 12000,
-	result: 12000,
-	details: 8000,
-};
-
-export interface AgentPanelGraphHeaderInput {
-	readonly title: string;
-	readonly subtitle?: string | null;
-	readonly agentIconSrc?: string | null;
-	readonly agentLabel?: string | null;
-	readonly projectLabel?: string | null;
-	readonly projectColor?: string | null;
-	readonly sequenceId?: number | null;
-	readonly branchLabel?: string | null;
-	readonly actions?: readonly AgentPanelActionDescriptor[];
-}
-
-export interface AgentPanelGraphMaterializerInput {
-	readonly panelId: string;
-	readonly graph: AgentPanelCanonicalSource | null;
-	readonly header: AgentPanelGraphHeaderInput;
-	readonly composer?: AgentPanelComposerModel | null;
-	readonly strips?: readonly AgentPanelStripModel[];
-	readonly cards?: readonly AgentPanelCardModel[];
-	readonly sidebars?: AgentPanelSidebarModel | null;
-	readonly chrome?: AgentPanelChromeModel | null;
-	readonly optimistic?: {
-		readonly pendingUserEntry?: SessionEntry | null;
-	} | null;
-}
 
 interface OperationIndex {
 	readonly byOperationId: Map<string, OperationSnapshot>;
@@ -127,10 +110,6 @@ interface CachedConversationState {
 		isStreaming: boolean;
 	};
 	readonly sceneEntryRowIndex: ReadonlyMap<string, number>;
-}
-
-export interface AgentPanelGraphMaterializerReadModel {
-	apply(input: AgentPanelGraphMaterializerInput): AgentPanelSceneModel;
 }
 
 function buildAssistantMessageFromTranscriptEntry(entry: TranscriptEntry) {
